+++ src/main/java/com/example/main/controllers/UserController.java
@@ -24,7 +24,7 @@ public class UserController {
     }
 
     @DeleteMapping("/users/{id}")
-    public void deleteOne(@PathVariable int id){
+    public void deleteOne(@PathVariable long id){
         userService.deleteOne(id);
     }
 }
