+++ src/main/java/com/example/main/repositories/LinkZaProveruRepository.java
@@ -3,5 +3,5 @@ package com.example.main.repositories;
 import com.example.main.models.LinkZaProveruModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LinkZaProveruRepository extends JpaRepository<LinkZaProveruModel, Integer> {
+public interface LinkZaProveruRepository extends JpaRepository<LinkZaProveruModel, Long> {
 }
