+++ src/main/java/com/example/main/repositories/RequestRepository.java
@@ -3,5 +3,5 @@ package com.example.main.repositories;
 import com.example.main.models.RequestModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RequestRepository extends JpaRepository<RequestModel, Integer> {
+public interface RequestRepository extends JpaRepository<RequestModel, Long> {
 }
